Completed -Added an Adapter for forecast items -Parsed the list to show in the adapter -Completed UI -Added location permission check -Handled error conditions

+++ app/src/main/java/com/example/weatherapp/DI/Models/Wind.kt
@@ -1,6 +1,6 @@
 package com.example.weatherapp.DI.Models
 
 data class Wind(
-    val deg: Double,
+    val deg: Int,
     val speed: Double
 )
